feat: move form to page
