+++ client-app/shared/account/components/index.ts
@@ -4,6 +4,7 @@ export { default as AccountNavigationLink } from "./account-navigation-link.vue"
 export { default as AccountShell } from "./account-shell.vue";
 export { default as ForgotPasswordForm } from "./forgot-password-form.vue";
 export { default as ResetPasswordForm } from "./reset-password-form.vue";
+export { default as ChangePasswordForm } from "./change-password-form.vue";
 export { default as RegistrationSuccessDialog } from "./registration-success-dialog.vue";
 export { default as CheckoutDefaultsSuccessDialog } from "./checkout-default-success-dialog.vue";
 export { default as ProfileUpdateSuccessDialog } from "./profile-update-success-dialog.vue";
