+++ locales/en.json
@@ -345,7 +345,6 @@
       },
       "change_password_form": {
         "title": "Create new password",
-        "header": "password change",
         "current_pass_label": "Current password",
         "current_pass_placeholder": "Enter your current password",
         "new_pass_label": "New password",
@@ -353,7 +352,9 @@
         "confirm_new_pass_label": "Confirm new password",
         "confirm_new_pass_placeholder": "Repeat your new password",
         "submit_btn": "create",
-        "wrong_current_pass": "You entered the wrong current password, please try again or call support"
+        "wrong_current_pass": "You entered the wrong current password, please try again or call support",
+        "logout": "logout",
+        "forgot_password": "Forgot your password?"
       },
       "sign_in_form": {
         "remember_me_label": "Remember me",
@@ -791,6 +792,14 @@
       "enter_new_password_message": "An account has been created for you, please set a password to login.",
       "success_message": "Registration was successfully completed",
       "log_in_button": "Continue"
+    },
+    "change_password": {
+      "meta": {
+        "title": "Change password"
+      },
+      "header": "Change your password",
+      "enter_new_password_message": "Enter your new password",
+      "success_message": "Your password has been successfully changed"
     },
     "sign_in": {
       "meta": {
